write run_endpoint.sh and haproxy config file
run_endpoint.sh is the entrypoint of the Docker container. It is written
to be conformant with QUIC interop simulator as described here :
https://github.com/marten-seemann/quic-network-simulator

This container is a valid server implementation based on haproxy. It
will also start a simple python http server to be able to serve HTML
files from the /www directory. The certificate and key generated by the
simulator are copied into /tmp and the key is renamed to respect the
naming convention used by haproxy.

When started as a client, the container will exit with status code 127
in conformance with the QUIC interop simulator.

The test "multiconnect" is not supported because at the moment it seems
to cause the simulator to hang.