lighttpd: allow GET requests with body
If the STREAM frame containing the H3 HEADERS frame has not the FIN bit
set, haproxy convert this as a chunked transfer-encoding body. This is
sometimes the case with clients such as s2n-quic.

By default lighttpd reject GET requests with body with the 400 HTTP
response status code. Edit lighttpd configuration to allow such requests
and ignore the body.