log ssl keys using a Lua script
Add a Lua script to log SSL keys to decipher traffic in an output file.
This is available through a http-request action. To execute only once
per connection a session variable is used.

To be able to use SSL keys fetch the setting tune.ssl.keylog must be
activated in haproxy configuration.

The output file is specified as a parameter to the http-request action.
It is set to /logs/keys.log which is the path inspected by the interop.

This will allow to support on the interop several tests which rely on
traffic deciphering even if the client do not log SSL keys. For the
moment, only mvfst seems to be in this case.