hap-cfg: add an alternative config file for backend testing
Add an alternative config file which will be used when haproxy acts as a
client in the context of QUIC interop.

This config is mostly identical to frontend testing, except of course
that QUIC is now used via server in the backend section. A simple
frontend listening on 20080 port is defined, which will be used to
transcode HTTP/1 requests into QUIC ones.

Experimental directives are activated as QUIC support on the backend is
not yet considered as stabled. Also, only hq-interop is currently used,
as HTTP/3 backend support has not yet been merged.

Also of notable interest, SSL keylog is not activated for backend
support for now. To be functional, it requires the usage of SSL fetches
dedicated to the backend side.