hap-cfg: prevent empty vars error
HAP_EXTRA_ARGS is an environment variable that serves to pass extra
arguments on haproxy startup. Currently, it is set to quic-force-retry
for retry test. In other cases it is an empty string.

A new error is reported on startup if a config variable is empty. To
prevent it in the interop, use the notation "var[*]" which is the
correct alternative.